tests: Merge src/tests/ into tests/
A lot is happening here, but I failed to find a way to break it
into smaller changes.

There are some tests under src/tests/ but we do not run them. So
the first step was to move them under tests/. Then, I've switched
individual tests from using runtests.sh (except for
test-domain-snapshot.phpt, which needs a special care and thus is
covered in next commit), to run-tests.php. This is a php test
suite and the script was copied over from php sources (just like
check_parameters.php was) [1]. This allows us to actually run
tests in parallel, as Makefile handles test execution, not a bash
script. But this means, that existing test cases, handled by
runtests.sh needed to be adjusted for new test suite [2]. In
here, a single file contains both, the test code and expected
output. It can even contain cleanup code.

After all of this, functions.phpt.in was basically useless and
the single useful piece of code left there was definition of
$abs_srcdir variable. And since it's included from other .phpt
files, it's renamed to functions.inc.in.

1: https://github.com/php/php-src/blob/master/run-tests.php
2: https://qa.php.net/phpt_details.php
Signed-off-by: Michal Privoznik <mprivozn@redhat.com>